+++ src/lib/index.ts
@@ -1,1 +1,2 @@
 export { Testing } from './testing.class';
+export { TestingToBeMatchers } from './testing-tobe-matchers.class';